UPD: split master's load + import in partial classes
